Move DatasetList to own file.
